+++ src/main/java/com/cryptocurrencies/api/domain/model/Wallet.java
@@ -6,11 +6,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Wallet {
     private long idUser;
     private String symbol;
-    private double funds;
+    private BigDecimal funds;
 }
